refactor: 보드 생성에 전략 패턴 도입
