+++ src/main/java/domain/board/BoardGenerator.java
@@ -1,0 +1,9 @@
+package domain.board;
+
+import domain.piece.Piece;
+import java.util.Map;
+
+public interface BoardGenerator {
+
+    Map<Position, Piece> generate();
+}
