+++ src/test/java/domain/board/InitialBoardGeneratorTest.java
@@ -1,0 +1,18 @@
+package domain.board;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class InitialBoardGeneratorTest {
+
+    @Test
+    @DisplayName("체스판을 초기화한다.")
+    void initialize() {
+        BoardGenerator boardGenerator = new InitialBoardGenerator();
+        Board board = Board.generatedBy(boardGenerator);
+
+        assertThat(board.getSquares()).hasSize(32); // TODO: 검증 로직 구체화
+    }
+}
